t2s_backend/pyproject.toml
[project]
name = "t2s-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "celery-types==0.23.0",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-unfold>=0.69.0",
    "djangorestframework>=3.16.1",
    "django-filter>=25.2",
    "drf-spectacular>=0.28.0",
    "uvicorn[standard]>=0.38.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "django-redis>=6.0.0",
    "sqlalchemy>=2.0.49",
    "transformers>=4.56.2",
    "torch>=2.5.1"

]
